# shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/webhooks/ads/{network}": {
      "post": {
        "operationId": "AdsWebhookController_handlePost",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ]
      },
      "get": {
        "operationId": "AdsWebhookController_handleGet",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ]
      }
    },
    "/api/notifications/register-token": {
      "post": {
        "operationId": "NotificationsController_registerToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/admin/gift-cards": {
      "get": {
        "operationId": "AdminGiftCardsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminGiftCardsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiftCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/gift-cards/{id}": {
      "patch": {
        "operationId": "AdminGiftCardsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiftCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/inventory": {
      "post": {
        "operationId": "AdminInventoryController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "AdminInventoryController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/inventory/{id}/reveal": {
      "get": {
        "operationId": "AdminInventoryController_reveal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/redemptions/{id}/approve": {
      "post": {
        "operationId": "AdminRedemptionsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/redemptions/{id}/reject": {
      "post": {
        "operationId": "AdminRedemptionsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRedemptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cash Raja API",
    "description": "Coin-rewards backend: auth, wallet/ledger, offers & ads postbacks, engagement (game/streak/scratch-spin/referral), redemptions, notifications, fraud, and the admin API.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "default"
    }
  ],
  "components": {
    "securitySchemes": {
      "user": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "App-user access token (aud=app)"
      },
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Admin access token (aud=admin)"
      }
    },
    "schemas": {
...
      "RegisterTokenDto": {
        "type": "object",
        "properties": {}
      },
      "CreateGiftCardDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateGiftCardDto": {
        "type": "object",
        "properties": {}
      },
      "InventoryUploadDto": {
        "type": "object",
        "properties": {}
      },
...
      "RejectRedemptionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}